# perf: high CPU usage while parsing concurrent flux queries