--- a/libflux/flux/Cargo.toml
+++ b/libflux/flux/Cargo.toml
@@ -45,3 +45,7 @@ harness = false
 [[bench]]
 name = "formatter"
 harness = false
+
+[[bench]]
+name = "analyze"
+harness = false