--- a/libflux/flux/src/cffi.rs
+++ b/libflux/flux/src/cffi.rs
@@ -25,7 +25,7 @@ use fluxcore::{
 
 use crate::semantic::flatbuffers::semantic_generated::fbsemantic::MonoTypeHolderArgs;
 
-use super::{imports, new_semantic_analyzer, prelude, Error, Result, IMPORTS, PRELUDE};
+use super::{new_semantic_analyzer, prelude, Error, Result, IMPORTS, PRELUDE};
 
 /// An error handle designed to allow passing `Error` instances to library
 /// consumers across language boundaries.
@@ -413,7 +413,7 @@ fn new_stateful_analyzer() -> Result<StatefulAnalyzer> {
         Some(prelude) => prelude,
         None => return Err(anyhow!("missing prelude").into()),
     };
-    let imports = match imports() {
+    let imports = match &*IMPORTS {
         Some(imports) => imports,
         None => return Err(anyhow!("missing stdlib imports").into()),
     };
@@ -424,13 +424,12 @@ fn new_stateful_analyzer() -> Result<StatefulAnalyzer> {
 /// This enables uses cases where analysis is performed iteratively, for example in a REPL.
 pub struct StatefulAnalyzer {
     env: PackageExports,
-    imports: Packages,
+    imports: &'static Packages,
 }
 
 impl StatefulAnalyzer {
     fn analyze(&mut self, ast_pkg: &ast::Package) -> Result<fluxcore::semantic::nodes::Package> {
-        let mut analyzer =
-            Analyzer::new_with_defaults(Environment::from(&self.env), mem::take(&mut self.imports));
+        let mut analyzer = Analyzer::new_with_defaults(Environment::from(&self.env), self.imports);
         let (mut env, sem_pkg) = match analyzer.analyze_ast(ast_pkg) {
             Ok(r) => r,
             Err(e) => {
@@ -600,11 +599,11 @@ fn find_var_type(ast_pkg: &ast::Package, var_name: String) -> Result<MonoType> {
 /// This function is unsafe because it dereferences a raw pointer.
 #[no_mangle]
 pub unsafe extern "C" fn flux_get_env_stdlib(buf: *mut flux_buffer_t) {
-    let imports = imports().unwrap();
+    let imports = IMPORTS.as_ref().unwrap();
     let env = PackageExports::try_from(
         imports
-            .into_iter()
-            .map(|(k, v)| (Symbol::from(k), v.typ()))
+            .iter()
+            .map(|(k, v)| (Symbol::from(k.as_str()), v.typ()))
             .collect::<Vec<_>>(),
     )
     .unwrap();